test: simplify SideNav integration testing
