--- vaadin-side-nav-flow-parent/vaadin-side-nav-flow-integration-tests/src/main/java/com/vaadin/flow/component/sidenav/tests/SideNavRouterLayoutTarget1Page.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2000-2023 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.flow.component.sidenav.tests;
-
-import com.vaadin.flow.component.html.Div;
-import com.vaadin.flow.component.html.Span;
-import com.vaadin.flow.router.Route;
-
-@Route(value = "vaadin-side-nav/side-nav-test-router-layout-target1", layout = SideNavRouterLayout.class)
-public class SideNavRouterLayoutTarget1Page extends Div {
-
-    public SideNavRouterLayoutTarget1Page() {
-        add(new Span("View to test routing from the vaadin-side-nav component"
-                + " - target 1"));
-    }
-}
